float field

+++ Appointment/models.py
@@ -98,10 +98,10 @@ class Appointment(models.Model):
     discount_type = models.CharField(max_length=50, choices= DISCOUNT_CHOICES, null=True, blank=True)
     payment_method = models.CharField(max_length=100, choices= PAYMENT_CHOICES, default='', null=True, blank=True)  
     
-    extra_price  = models.PositiveIntegerField(default=0, null=True, blank=True)
-    tip  = models.PositiveIntegerField(default=0, null=True, blank=True)
+    extra_price  = models.FloatField(default=0, null=True, blank=True)
+    tip  = models.FloatField(default=0, null=True, blank=True)
     
-    discount_price  = models.PositiveIntegerField(default=0, null=True, blank=True)
+    discount_price  = models.FloatField(default=0, null=True, blank=True)
     discount_percentage = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)
     
     service_commission = models.PositiveBigIntegerField(default = 0 , null=True, blank=True)    
@@ -162,7 +162,7 @@ class AppointmentService(models.Model):
     slot_availible_for_online = models.CharField(max_length=100, default='', null=True, blank=True,)
     
     appointment_status = models.CharField(choices=BOOKED_CHOICES, max_length=100, default='Appointment Booked')
-    tip = models.PositiveIntegerField(default=0, null=True, blank=True)
+    tip = models.FloatField(default=0, null=True, blank=True)
     
     price = models.FloatField(default=0, null=True, blank=True)
     
@@ -237,7 +237,7 @@ class AppointmentCheckout(models.Model):
     membership =models.ForeignKey(Membership, on_delete=models.CASCADE, related_name='checkout_membership_appointments', null=True, blank=True) 
     rewards =models.ForeignKey(Rewards, on_delete=models.CASCADE, related_name='checkout_reward_appointments', null=True, blank=True) 
     
-    tip = models.PositiveIntegerField(default=0, null=True, blank=True) # this field is not in use
+    tip = models.FloatField(default=0, null=True, blank=True) # this field is not in use
     gst = models.FloatField(default=0, null=True, blank=True)
     gst_price = models.FloatField(default=0, null=True, blank=True)
